+++ Changes.md
@@ -1,3 +1,11 @@
+## 0.0.4 - 2024-08-16
+
+- Added `time current` and `time context` commands
+
+## 0.0.3 - 2024-08-16
+
+- Added `dnd status` and `dnd time` commands
+
 ## 0.0.2 - 2024-07-22
 
 - Changed the endpoint to one.thing.tools